Refactor in load data
