+++ src/main.py
@@ -6,51 +6,30 @@ from flask import Flask, request, jsonify, url_for
 from flask_migrate import Migrate
 from flask_swagger import swagger
 from flask_cors import CORS
+
 from utils import APIException, generate_sitemap
 from admin import setup_admin
-from models import db, User
-from create_database import create_db
-from seed_data import seed_data
-
-#from models import Person
+from models import db
+from init_database import create_db, load_seed_data
+from seed_data import data
 
 app = Flask(__name__)
-data_base = os.environ.get('DB_CONNECTION_STRING')
-print(data_base)
+app.app_context().push()
+data_base = os.environ['DB_CONNECTION_STRING']
 create_db(data_base)
-
 
 app.url_map.strict_slashes = False
 app.config['SQLALCHEMY_DATABASE_URI'] = data_base
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
 MIGRATE = Migrate(app, db)
-# aqui se crean las tablas 
 db.init_app(app)
+load_seed_data(data)
+
 CORS(app)
 setup_admin(app)
-seed_data()
 
 
-# Handle/serialize errors like a JSON object
-@app.errorhandler(APIException)
-def handle_invalid_usage(error):
-    return jsonify(error.to_dict()), error.status_code
-
-# generate sitemap with all your endpoints
-@app.route('/')
-def sitemap():
-    return generate_sitemap(app)
-
-@app.route('/user', methods=['GET'])
-def handle_hello():
-
-    response_body = {
-        "msg": "Hello, this is your GET /user response "
-    }
-
-    return jsonify(response_body), 200
-
-# this only runs if `$ python src/main.py` is executed
 if __name__ == '__main__':
     PORT = int(os.environ.get('PORT', 3000))
     app.run(host='0.0.0.0', port=PORT, debug=False)
